Every day it's getting closer, moving faster than a roller coaster

+++ Examples/_7:SimpleLSVR.py
@@ -25,83 +25,83 @@ if __name__ == "__main__":
     Themecolors().GnuChanOS
 
     win = [
-        [   gc.GText(title="This Is First Version gpu-screen-recorder GUI", xStretch=True, position="center", font=gc.font)   ],
-        [   gc.GText(title="Video or LiveStream Time", xStretch=True, position="center", bcolor=GColors().purple8, font=gc.font)   ],
-        [   gc.GText(title="0:0:0", value="time", xStretch=True, position="center", font="Sans, 20")   ],
+        [   gc.GText(SetText="This Is First Version gpu-screen-recorder GUI", xStretch=True, TPosition="center", TFont=gc.font)   ],
+        [   gc.GText(SetText="Video or LiveStream Time", xStretch=True, TPosition="center", BColor=GColors().purple8, TFont=gc.font)   ],
+        [   gc.GText(SetText="0:0:0", TextValue="time", xStretch=True, TPosition="center", TFont="Sans, 20")   ],
 
-        [   gc.GText(title="Video FPS", xStretch=True, position="center", bcolor=GColors().purple8, font=gc.font)   ],
+        [   gc.GText(SetText="Video FPS", xStretch=True, TPosition="center", BColor=GColors().purple8, TFont=gc.font)   ],
         [
-            gc.Push(bcolor=GnuChanOSColor().colors1),
-            gc.GRadio(value="14", groupID="fps", title="14-FPS", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="24", groupID="fps", title="24-FPS", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="30", groupID="fps", title="30-FPS", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="60", groupID="fps", title="60-FPS", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="120", groupID="fps", title="120-FPS", bcolor=GColors().purple7, font=gc.font),
-            gc.Push(bcolor=GnuChanOSColor().colors1),
-        ],
-
-        [   gc.GText(title="Sound Quality", xStretch=True, position="center", bcolor=GColors().purple8, font=gc.font)   ],
-        [
-            gc.Push(bcolor=GnuChanOSColor().colors1),
-            gc.GRadio(value="128", groupID="Kbps", title="128 Kbps", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="192", groupID="Kbps", title="192 Kbps", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="256", groupID="Kbps", title="256 Kbps", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="320", groupID="Kbps", title="320 Kbps", bcolor=GColors().purple7, font=gc.font),
-            gc.Push(bcolor=GnuChanOSColor().colors1),
+            gc.GPush(BColor=GnuChanOSColor().colors1),
+            gc.GRadio(WindowValue="14", groupID="fps", RText="14-FPS", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="24", groupID="fps", RText="24-FPS", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="30", groupID="fps", RText="30-FPS", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="60", groupID="fps", RText="60-FPS", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="120", groupID="fps", RText="120-FPS", BColor=GColors().purple7, TFont=gc.font),
+            gc.GPush(BColor=GnuChanOSColor().colors1),
         ],
 
-        [   gc.GText(title="Video Quality", xStretch=True, position="center", bcolor=GColors().purple8, font=gc.font)   ],
+        [   gc.GText(SetText="Sound Quality", xStretch=True, TPosition="center", BColor=GColors().purple8, TFont=gc.font)   ],
         [
-            gc.Push(bcolor=GnuChanOSColor().colors1),
-            gc.GRadio(value="Medium", groupID="quality", title="Medium", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="high", groupID="quality", title="High", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="very_high", groupID="quality", title="Very High", bcolor=GColors().purple7, font=gc.font),
-            gc.GRadio(value="ultra", groupID="quality", title="Ultra", bcolor=GColors().purple7, font=gc.font),
-            gc.Push(bcolor=GnuChanOSColor().colors1),
+            gc.GPush(BColor=GnuChanOSColor().colors1),
+            gc.GRadio(WindowValue="128", groupID="Kbps", RText="128 Kbps", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="192", groupID="Kbps", RText="192 Kbps", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="256", groupID="Kbps", RText="256 Kbps", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="320", groupID="Kbps", RText="320 Kbps", BColor=GColors().purple7, TFont=gc.font),
+            gc.GPush(BColor=GnuChanOSColor().colors1),
         ],
-        [   gc.GText(title="Monutor ID 'xrandr here': ", font=gc.font),
-            gc.GInput(value="mID", xStretch=True, size=(20, None), font=gc.font)
+
+        [   gc.GText(SetText="Video Quality", xStretch=True, TPosition="center", BColor=GColors().purple8, TFont=gc.font)   ],
+        [
+            gc.GPush(BColor=GnuChanOSColor().colors1),
+            gc.GRadio(WindowValue="Medium", groupID="quality", RText="Medium", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="high", groupID="quality", RText="High", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="very_high", groupID="quality", RText="Very High", BColor=GColors().purple7, TFont=gc.font),
+            gc.GRadio(WindowValue="ultra", groupID="quality", RText="Ultra", BColor=GColors().purple7, TFont=gc.font),
+            gc.GPush(BColor=GnuChanOSColor().colors1),
+        ],
+        [   gc.GText(SetText="Monutor ID 'xrandr here': ", TFont=gc.font),
+            gc.GInput(WindowValue="mID", xStretch=True, Size=(20, None), TFont=gc.font)
         ],
         [
-            gc.GText(title="Micraphone Device ID: ", font="Sans, 15"), gc.GInput(value="mic", xStretch=True, size=(20, None), font="Sans, 15"),
+            gc.GText(SetText="Micraphone Device ID: ", TFont="Sans, 15"), gc.GInput(WindowValue="mic", xStretch=True, Size=(20, None), TFont="Sans, 15"),
         ],
         [
-            gc.GText(title="Desktop Sound Device ID: ", font="Sans, 15"), gc.GInput(value="desk", xStretch=True, size=(20, None), font="Sans, 15"),
+            gc.GText(SetText="Desktop Sound Device ID: ", TFont="Sans, 15"), gc.GInput(WindowValue="desk", xStretch=True, Size=(20, None), TFont="Sans, 15"),
         ],
-        [   gc.GText(title="Live Stream Settings", xStretch=True, position="center", bcolor=GColors().purple8, font=gc.font)   ],
-        [   gc.GText(title="RTMP URL: ", font=gc.font),
-            gc.GInput(value="rtmp", xStretch=True, size=(20, None), font=gc.font)
+        [   gc.GText(SetText="Live Stream Settings", xStretch=True, TPosition="center", BColor=GColors().purple8, TFont=gc.font)   ],
+        [   gc.GText(SetText="RTMP URL: ", TFont=gc.font),
+            gc.GInput(WindowValue="rtmp", xStretch=True, Size=(20, None), TFont=gc.font)
         ],
-        [   gc.GText(title="Stream key: ", font=gc.font),
-            gc.GInput(value="skey", xStretch=True, PwChars="*", size=(20, None), font=gc.font)
+        [   gc.GText(SetText="Stream key: ", TFont=gc.font),
+            gc.GInput(WindowValue="skey", xStretch=True, HidePassword="*", Size=(20, None), TFont=gc.font)
         ],
         [   
-            gc.Push(bcolor=GnuChanOSColor().colors1),
-            gc.GButton(title="Save And Start Live Stream", font=gc.font), gc.GButton(title="Stop Live Stream", font=gc.font),
-            gc.Push(bcolor=GnuChanOSColor().colors1),
+            gc.GPush(BColor=GnuChanOSColor().colors1),
+            gc.GButton(Text="Save And Start Live Stream", TFont=gc.font), gc.GButton(Text="Stop Live Stream", TFont=gc.font),
+            gc.GPush(BColor=GnuChanOSColor().colors1),
         ],
-        [   gc.GText(title="Video Record", xStretch=True, position="center", bcolor=GColors().purple8, font=gc.font)   ],
-        [   gc.GText(title="File Path Here!", value="path", xStretch=True, bcolor=GColors().purple7, font=gc.font)],
-        [   gc.GText(title="Video Name: ", font=gc.font), gc.GInput(value="vname", size=(20, None), xStretch=True, font=gc.font)   ],
+        [   gc.GText(SetText="Video Record", xStretch=True, TPosition="center", BColor=GColors().purple8, TFont=gc.font)   ],
+        [   gc.GText(SetText="File Path Here!", TextValue="path", xStretch=True, BColor=GColors().purple7, TFont=gc.font)],
+        [   gc.GText(SetText="Video Name: ", TFont=gc.font), gc.GInput(WindowValue="vname", Size=(20, None), xStretch=True, TFont=gc.font)   ],
         [
-            gc.Push(bcolor=GnuChanOSColor().colors1),
-                gc.GButton(title="Select Output Folder", font=gc.font),
-                gc.GButton(title="Save And Start Record", font=gc.font),
-                gc.GButton(title="Stop Record", font=gc.font),
-            gc.Push(bcolor=GnuChanOSColor().colors1),
+            gc.GPush(BColor=GnuChanOSColor().colors1),
+                gc.GButton(Text="Select Output Folder", TFont=gc.font),
+                gc.GButton(Text="Save And Start Record", TFont=gc.font),
+                gc.GButton(Text="Stop Record", TFont=gc.font),
+            gc.GPush(BColor=GnuChanOSColor().colors1),
         ]
 
     ]
 
     layout = [ 
         [   
-            gc.vsep(color=GColors().black),
-            gc.GColumn(winColumn=win, xStretch=True, yStretch=True),
-            gc.vsep(color=GColors().black),
+            gc.GVSep(Color=GColors().black),
+            gc.GColumn(winColumnLayout_List=win, xStretch=True, yStretch=True),
+            gc.GVSep(Color=GColors().black),
         ]
             ]
 
-    gc.GWindow(mainWindow=layout)
+    gc.GWindow(SetMainWindowLayout_List=layout)
 
     _rtmp = _key = _monutorID = _fps = _quality = _sound = ""
     _path = _videoName = ""
@@ -109,21 +109,22 @@ if __name__ == "__main__":
     _desktop = "$(pactl get-default-sink).monitor"
     _mic = "alsa_input.usb-Generic_USB2.0_PC_CAMERA_20100331010203-02.mono-fallback"
     _now = datetime.now()
+    _LiveOrRecordStart = False
 
-    gc.window["rtmp"].update("rtmp://live.restream.io/live")
-    gc.window["mID"].update("VGA1")
-    gc.window["mic"].update(_mic)
-    gc.window["desk"].update(_desktop)
+    gc.GetWindow["rtmp"].update("rtmp://a.rtmp.youtube.com/live2")
+    gc.GetWindow["mID"].update("VGA1")
+    gc.GetWindow["mic"].update(_mic)
+    gc.GetWindow["desk"].update(_desktop)
 
     def LiveStream():
-        global _rtmp, _key, _monutorID, _fps, _quality, _sound, _path, _videoName, _Start, _now, _mic, _desktop
-        if gc.event in ["14", "24", "30", "60", "120"]:
-            _fps = gc.event
-        if gc.event in ["Medium", "high", "very_high", "ultra"]:
-            _quality = str(gc.event).lower()
-        if gc.event in ["128", "192", "256", "320"]:
-            _sound = gc.event
-        elif gc.event == "Save And Start Live Stream":
+        global _rtmp, _key, _monutorID, _fps, _quality, _sound, _path, _videoName, _Start, _now, _mic, _desktop, _LiveOrRecordStart
+        if gc.GetEvent in ["14", "24", "30", "60", "120"]:
+            _fps = gc.GetEvent
+        if gc.GetEvent in ["Medium", "high", "very_high", "ultra"]:
+            _quality = str(gc.GetEvent).lower()
+        if gc.GetEvent in ["128", "192", "256", "320"]:
+            _sound = gc.GetEvent
+        elif gc.GetEvent == "Save And Start Live Stream":
             _rtmp = gc.GetValues["rtmp"]
             _key = gc.GetValues["skey"]
             _monutorID = gc.GetValues["mID"]
@@ -140,25 +141,27 @@ if __name__ == "__main__":
                         os.popen(_liveStream)
                         os.popen("notify-send -t 3500 -u low \"Live Stream is Starting!\"")
                         _Start = True
+                        _LiveOrRecordStart = True
             except Exception as ERR:
                 print(ERR)
 
-        elif gc.event == "Stop Live Stream":
+        elif gc.GetEvent == "Stop Live Stream":
             os.popen("killall -SIGINT gpu-screen-recorder && notify-send -t 3500 -u low \"Work Is Finish!\"")
             _Start = False
+            _LiveOrRecordStart = False
 
     def RecordVideo():
         global _monutorID, _fps, _quality, _sound, _videoName, _Start, _path, _desktop, _mic
         _monutorID = gc.GetValues["mID"]
         _videoName = gc.GetValues["vname"]
-        if gc.event in ["14", "24", "30", "60", "120"]:
-            _fps = gc.event
-        if gc.event in ["Medium", "high", "very_high", "ultra"]:
-            _quality = str(gc.event).lower()
-        if gc.event in ["128", "192", "256", "320"]:
-            _sound = gc.event
+        if gc.GetEvent in ["14", "24", "30", "60", "120"]:
+            _fps = gc.GetEvent
+        if gc.GetEvent in ["Medium", "high", "very_high", "ultra"]:
+            _quality = str(gc.GetEvent).lower()
+        if gc.GetEvent in ["128", "192", "256", "320"]:
+            _sound = gc.GetEvent
 
-        if gc.event == "Save And Start Record":
+        if gc.GetEvent == "Save And Start Record":
             if _path and _videoName and _monutorID and _fps and _quality and _sound != "":
                 if not _Start:
                     # change this place for your device
@@ -172,19 +175,21 @@ if __name__ == "__main__":
                     os.popen(_VideoRecord)
                     os.popen("notify-send -t 3500 -u low \"Video Recording is Starting\"")
                     _Start = True
+                    _LiveOrRecordStart = True
 
-        elif gc.event == "Stop Record":
+        elif gc.GetEvent == "Stop Record":
             os.popen("killall -SIGINT gpu-screen-recorder && notify-send -t 3500 -u low \"Work Is Finish!\"")
             _Start = False
+            _LiveOrRecordStart = False
 
     _second = _minute = _hour = 0
     _sec = time.monotonic()
 
     def update():
         global _path
-        if gc.event == "Select Output Folder":
+        if gc.GetEvent == "Select Output Folder":
             _path = popup_get_folder(no_window=True, title="Warning!", message="Select Output Folder For Recording Video")
-            gc.window["path"].update(_path)
+            gc.GetWindow["path"].update(_path)
             print(_path)
 
         Thread(target=LiveStream, args=[]).start()
@@ -203,12 +208,13 @@ if __name__ == "__main__":
                 _second = 0
         else:
             _second = _minute = _hour = 0
-        gc.window["time"].update(f"{_hour}:{_minute}:{int(_second)}")
+        gc.GetWindow["time"].update(f"{_hour}:{_minute}:{int(_second)}")
 
     def LastWorkBeforeExit():
-        global _Start
+        global _Start, _LiveOrRecordStart
         if _sound:
-            os.popen("killall -SIGINT gpu-screen-recorder && notify-send -t 3500 -u low \"Work Is Finish!\"")
-            _Start = False
+            if _LiveOrRecordStart:
+                os.popen("killall -SIGINT gpu-screen-recorder && notify-send -t 3500 -u low \"Work Is Finish!\"")
+                _Start = False
 
-    gc.update(GUpdate=update, exitBEFORE=LastWorkBeforeExit)
+    gc.SetUpdate(Update=update, exitBEFORE=LastWorkBeforeExit)
